AI-Assisted Diagnostic Reporting Support/backend/ai_engine/vit_inference.py
import torch
from transformers import AutoFeatureExtractor, AutoModelForImageClassification
from PIL import Image
import numpy as np

# Using a public ViT fine-tuned on chest X-rays (example)
MODEL_NAME = "google/vit-base-patch16-224-in21k"  # Replace with fine-tuned if available

feature_extractor = AutoFeatureExtractor.from_pretrained(MODEL_NAME)
model = AutoModelForImageClassification.from_pretrained(MODEL_NAME)
model.eval()

def get_vit_heatmap(image_array: np.ndarray):
    """Returns dummy heatmap coordinates – replace with real Grad-CAM/attention rollout"""
    pil_img = Image.fromarray(image_array)
    inputs = feature_extractor(images=pil_img, return_tensors="pt")

    with torch.no_grad():
        outputs = model(**inputs, output_attentions=True)

    # Dummy heatmap: center highlight
    h, w = image_array.shape[:2]
    heatmap = np.zeros((h, w), dtype=np.float32)
    heatmap[h//4:3*h//4, w//4:3*w//4] = 1.0

    findings = "Possible consolidation in central lung fields."
    return heatmap, findings