AI-Assisted Diagnostic Reporting Support/requirements.txt.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.32.0
pydicom==2.4.4
numpy==1.26.2
opencv-python==4.9.0.80
torch==2.3.0
transformers==4.38.2
pytesseract==0.3.10
pillow==10.2.0
python-multipart==0.0.6
pyyaml==6.0.1
pytest==8.0.0